V0.61
- Support uploads from AWTRIX flows
- Allows to send multiple custompages in one json
- Adds moodlight api
- Slightly better gif loading performance
- Power API now handles json OR true/false OR 0/1, also  Sound API can use json OR strings

+++ src/MatrixDisplayUi.cpp
@@ -79,9 +79,9 @@ void MatrixDisplayUi::setsetAutoTransitionBackwards()
   this->state.appTransitionDirection = -1;
   this->lastTransitionDirection = -1;
 }
-void MatrixDisplayUi::setTimePerApp(uint16_t time)
+void MatrixDisplayUi::setTimePerApp(long time)
 {
-  this->ticksPerApp = (int)((float)time / (float)updateInterval);
+  this->ticksPerApp = time / (long)updateInterval;
 }
 void MatrixDisplayUi::setTimePerTransition(uint16_t time)
 {
